requirements.txt
Django==4.2.19
djangorestframework==3.14.0
drf-yasg==1.21.4
mysqlclient==2.1.1
gunicorn==20.1.0